% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comparison}
\alias{comparison}
\title{GP Practice Comparison Data}
\format{
A data frame that contains GP practice data relating to practice and
patient sociodemographics, practice location, and practice workforce
}
\usage{
comparison
}
\description{
A dataset for building organisation comparison/similarity models using GP
practice data
}
\keyword{datasets}
