% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fft}
\alias{fft}
\title{Friends & Family Test Survey Data}
\format{
A data frame that contains GP practice, PCN, ICB, and region
identifiers, the survey month, and responses
}
\usage{
fft
}
\description{
Friends & Family Test survey responses aggregated to GP Practice level.
}
\keyword{datasets}
