% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gpps}
\alias{gpps}
\title{GP Patient Survey Data}
\format{
A data frame that contains GP practice, PCN, ICB, and region
identifiers, the survey year, survey question number and description, and
responses
}
\usage{
gpps
}
\description{
GP Patient Survey responses aggregated to GP Practice level.
}
\keyword{datasets}
