% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ods}
\alias{ods}
\title{Organisation Data Service (ODS) Data}
\format{
A data frame that contains GP practice name and code, PCN name and
code, ICB name and code, and region name and code.
}
\usage{
ods
}
\description{
A dataset for identifying organisations across the NHS. Includes GP practice
, PCN, ICB, and region codes.
}
\keyword{datasets}
